+++ emmet-builders/emmet/builders/materials/optimade.py
@@ -1,37 +1,36 @@
 from maggma.builders.map_builder import MapBuilder
 from maggma.core import Store
-from pymatgen.core import Structure
-from pymatgen.core import __version__ as pymatgen_version
+from pymatgen.core.structure import Structure
 
-from emmet.core.oxidation_states import OxidationStateDoc
+from emmet.core.optimade import OptimadeMaterialsDoc
 from emmet.core.utils import jsanitize
 
 
-class OxidationStatesBuilder(MapBuilder):
+class OptimadeMaterialsBuilder(MapBuilder):
     def __init__(
         self,
         materials: Store,
-        oxidation_states: Store,
+        optimade: Store,
         **kwargs,
     ):
         """
-        Creates Oxidation State documents from materials
+        Creates Optimade compatible structure docs for the materials
 
         Args:
             materials: Store of materials docs
-            oxidation_states: Store to update with oxidation state document
+            optimade: Store to update with optimade document
             query : query on materials to limit search
         """
         self.materials = materials
-        self.oxidation_states = oxidation_states
+        self.optimade = optimade
         self.kwargs = kwargs
 
         # Enforce that we key on material_id
         self.materials.key = "material_id"
-        self.oxidation_states.key = "material_id"
+        self.optimade.key = "material_id"
         super().__init__(
             source=materials,
-            target=oxidation_states,
+            target=optimade,
             projection=["structure"],
             **kwargs,
         )
@@ -39,10 +38,11 @@ class OxidationStatesBuilder(MapBuilder):
     def unary_function(self, item):
         structure = Structure.from_dict(item["structure"])
         mpid = item["material_id"]
+        last_updated = item["last_updated"]
 
-        oxi_doc = OxidationStateDoc.from_structure(
-            structure=structure, material_id=mpid
+        optimade_doc = OptimadeMaterialsDoc.from_structure(
+            structure=structure, material_id=mpid, last_updated=last_updated
         )
-        doc = jsanitize(oxi_doc.dict(), allow_bson=True)
+        doc = jsanitize(optimade_doc.dict(), allow_bson=True)
 
         return doc