+++ emmet-builders/emmet/builders/vasp/materials.py
@@ -8,6 +8,7 @@ from maggma.utils import grouper
 
 from pymatgen.core import Structure
 from emmet.builders.settings import EmmetBuildSettings
+from emmet.builders import SETTINGS
 from emmet.core.utils import group_structures, jsanitize
 from emmet.core.vasp.material import MaterialsDoc
 from emmet.core.vasp.task import TaskDocument
@@ -176,13 +177,18 @@ class MaterialsBuilder(Builder):
             "task_id",
             "formula_pretty",
             "output.energy_per_atom",
-            "output.energy",
             "output.structure",
+            "input.parameters",
+            # needed for run_type and task_type
             "calcs_reversed.input.parameters",
+            "calcs_reversed.input.incar",
+            "orig_inputs",
+            # needed for entry from task_doc
+            "output.energy",
             "input.is_hubbard",
             "input.hubbards",
             "input.potcar_spec",
-            "orig_inputs",
+            # misc info for materials doc
             "input.structure",
             "tags",
         ]
@@ -218,13 +224,23 @@ class MaterialsBuilder(Builder):
         materials = []
         for group in grouped_tasks:
             try:
-                materials.append(MaterialsDoc.from_tasks(group))
-            except Exception:
-                failed_ids = list({t_.task_id for t_ in group})
-                self.logger.warn(
-                    f"No valid ids found among ids {failed_ids}. This can be the case if the required "
-                    "calculation types are missing from your tasks database."
+                materials.append(
+                    MaterialsDoc.from_tasks(
+                        group,
+                        quality_scores=self.settings.VASP_QUALITY_SCORES,
+                        use_statics=self.settings.VASP_USE_STATICS,
+                    )
                 )
+            except Exception as e:
+                failed_ids = list({t_.task_id for t_ in group})
+                doc = MaterialsDoc.construct_deprecated_material(tasks)
+                doc.warnings.append(str(e))
+                materials.append(doc)
+                self.logger.warn(
+                    f"Failed making material for {failed_ids}."
+                    f" Inserted as deprecated Material: {doc.material_id}"
+                )
+
         self.logger.debug(f"Produced {len(materials)} materials for {formula}")
 
         return jsanitize([mat.dict() for mat in materials], allow_bson=True)